// src/vendor/ikoabo/models/errors.enum.ts
/**
 * Copyright (C) 2020 IKOA Business Opportunity
 * All Rights Reserved
 * Author: Reinier Millo Sánchez <reinier.millo88@gmail.com>
 *
 * This file is part of the IKOA Business Opportunity Core Server.
 * It can't be copied and/or distributed without the express
 * permission of the author.
 */

 export enum ERRORS {
   INVALID_OPERATION = 701,
   OBJECT_NOT_FOUND = 702,
   OBJECT_DUPLICATED = 703,
   INVALID_FIELDS = 704,
 }