// src/vendor/ikoabo/models/status.enum.ts
/**
 * Copyright (C) 2020 IKOA Business Opportunity
 * All Rights Reserved
 * Author: Reinier Millo Sánchez <reinier.millo88@gmail.com>
 *
 * This file is part of the IKOA Business Opportunity Core Server.
 * It can't be copied and/or distributed without the express
 * permission of the author.
 */

 export enum BASE_STATUS {
  BS_SOFT_DELETE = -1,
  BS_UNKNOWN = 0,
  BS_DISABLED = 1,
  BS_ENABLED = 2,
}